+++ DevelopersGuide.txt
@@ -17,7 +17,6 @@ python3 setup.py sdist bdist_wheel
 twine upload dist/*
 
 
-
 ### Type of commits:
 
 ENH: Enhancement, new functionality
